requirements.txt
# Core Dependencies
flask==2.3.3
flask-cors==4.0.0
flask-caching==2.1.0
flask-limiter==3.5.0
flask-jwt-extended==4.5.3

# Data Science
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
redis==5.0.0

# Development
pytest==7.4.0
black==23.7.0
flake8==6.1.0